add all describe api tests suits in one unique file api.test.js
